refactor contacts list into a hook
