+++ src/hooks/index.js
@@ -2,6 +2,7 @@ import useField from './useField';
 import useSession from './useSession';
 import useDataset from './useDataset';
 import useStatusNotification from './useStatusNotification';
+import useContactsList from './useContactsList';
 /**
  * The hooks module contains custom hooks to assist with form handling or status
  * notifications
@@ -9,4 +10,4 @@ import useStatusNotification from './useStatusNotification';
  * @namespace hooks
  */
 
-export { useField, useStatusNotification, useSession, useDataset };
+export { useField, useStatusNotification, useSession, useDataset, useContactsList };
